+++ Event/DynFormSavedEvent.php
@@ -2,6 +2,7 @@
 
 namespace L91\Sulu\Bundle\FormBundle\Event;
 
+use L91\Sulu\Bundle\FormBundle\Entity\Dynamic;
 use Symfony\Component\EventDispatcher\Event;
 
 class DynFormSavedEvent extends Event
@@ -11,25 +12,54 @@ class DynFormSavedEvent extends Event
     /**
      * @var array
      */
-    protected $formSelect;
+    protected $data;
+
+    /**
+     * @var Dynamic
+     */
+    protected $dynamic;
 
     /**
      * DynFormSavedEvent constructor.
      *
-     * @param $formSelect
+     * @param array $data
+     * @param Dynamic $dynamic will be required in the future
      */
-    public function __construct($formSelect)
+    public function __construct($data, $dynamic = null)
     {
-        $this->formSelect = $formSelect;
+        $this->data = $data;
+        $this->dynamic = $dynamic;
     }
 
     /**
      * Get FormSelect.
      *
+     * @deprecated use the getData function
+     *
      * @return array
      */
     public function getFormSelect()
     {
-        return $this->formSelect;
+        return $this->data;
+    }
+
+    /**
+     * Get data.
+     *
+     * @return array
+     */
+    public function getData()
+    {
+        return $this->data;
+    }
+
+    /**
+     * Get dynamic.
+     *
+     * @return Dynamic|null
+     */
+    public function getDynamic()
+    {
+        return $this->dynamic;
     }
 }